move cancel_trip_booking to trip_service to avoid circular import
